Fixed #73
